+++ migrations/20260313031506_event_invitation.js
@@ -5,8 +5,8 @@
 exports.up = function(knex) {
     return knex.schema.createTable('event_invitation', table => {
         table.increments();
-        table.integer('event_id').references('event.id');
-        table.integer('user_id').references('user.id');
+        table.bigint('event_id').references('event.id');
+        table.bigint('user_id').references('user.id');
         table.unique(['event_id', 'user_id'], {
             indexName: 'event_invitation_composite_index',
             useConstraint: true,
